requirements.txt
flask
scikit-learn
numpy
pandas
python-docx
pymupdf
